+++ lib/src/flutter/utils/driver_utils.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter_driver/flutter_driver.dart';
 
 class FlutterDriverUtils {
@@ -49,5 +51,32 @@ class FlutterDriverUtils {
       {Duration timeout = const Duration(seconds: 30)}) async {
     await driver.tap(finder, timeout: timeout);
     await FlutterDriverUtils.waitForFlutter(driver, timeout: timeout);
+  }
+
+  /// Waits until the [condition] returns true
+  /// Will raise a complete with a [TimeoutException] if the
+  /// condition does not return true with the timeout period.
+  static Future<void> waitUntil(
+    FlutterDriver driver,
+    Future<bool> condition(), {
+    Duration timeout = const Duration(seconds: 10),
+    Duration pollInterval = const Duration(milliseconds: 500),
+  }) async {
+    return Future.microtask(() async {
+      final completer = Completer<void>();
+      int maxAttempts =
+          (timeout.inMilliseconds / pollInterval.inMilliseconds).round();
+      int attempts = 0;
+
+      while (attempts < maxAttempts) {
+        final result = await condition();
+        if (result) {
+          completer.complete();
+          break;
+        } else {
+          await Future.delayed(pollInterval);
+        }
+      }
+    }).timeout(timeout);
   }
 }